Removed api.go, added Print interface, begun character building in main
The api directory was refactored into the fetch directory, with api.go
being removed. In addition to this, raceFetch.go was added, following a
similar format to the other fetch go files.
The contents of api.go were
moved to reference.go, along with adding a Printer interface.

Main now includes parsing for command line args, which is used for
building a path to the user-provided .toml file.

The buildCharacter function was added as well, which takes in references
to the necessary structs and returns a reference to the Character
struct.

The next update will be focused on improving functionality in main to
include actually fetching and unmarshalling data from the 5e wiki based
on the provided .toml file