Added fetchJSON and modified buildCharacter; GetEndpoint changed
fetchJSON takes a url string, target struct, waitgroup, and error
channel. It does an http GET request to the url and checks for errors.
If there are no errors, the request is unmarsheled into the provided
struct.

buildCharacter was modified to use goroutines to concurrently fetch
character properties using fetchJSON, then builds a Character struct
using references to those structs.

The GetEndpoint method has been changed to just get the endpoint for the
general character property (e.g. races/, classes/, etc) instead of the
endpoint for a specific race/class/etc. This makes it easier to
dynamically get the endpoint for any Fetchable struct.

The next update will be focused on making the TOML structure more
robust, and allowing for multiple spells/items (equipment).