Added empty TOML generator, removed tracking on JSON & TOML files
The templates/generator.go file now generates an empty TOML file that
matches the Go struct the characters are parsed into. The generator will
be updated alongside the template structs to ensure characters are
consistent.

In addition, characters/ and template/toml-characters/ have been added
to the .gitignore and will no longer be part of the commits.

Next update will be focused on adding more character features, such as
proficiency bonus choices.