Added support for Racial bonus to Ability Scores
In buildAbilityScores, it now takes in the race struct and uses it to
add the racial bonuses to the base template ability score. This means
the new value will now correctly be used when calculating the skill
bonuses.

Next update will focus on adding Combat Stats (HP, Temp HP, AC, etc)
with the struct model file having already been created.