REFACTOR: Major refactor of project structure from model/builder to DDD
All files within models/ and builder/ were moved into domain-specific
directories within the new internal/ directory.

Specific notes:
reference_type.go was moved into a reference/ directory and renamed to
type.go

The core/ directory was made to hold types and functions that are used
widely across the codebase. This includes:
- The ability enum, which several other types rely on (which caused cyclical imports)
- The FetchJSON function, which is used by race, class, spell, and
  inventory builder functions

The models and builder directories were removed and all import paths and
package names were updated. There may be smaller refactors necessary
down the line to keep the codebase clean and functional.

For the time being, the next update will introduce the Cobra CLI
package, which for now will be for saving, loading, and printing
characters from the TOML-to-JSON process.
In the future, it will also be used for launching tview.