refactor: Updated the fetch function and added fetcher interface
fetchJSONWithClient has been renamed to fetch.go. The core logic is the same, with
it using an input http.Client and a package level constant
DefaultBaseURL for the API.

fetcher.go defines the fetcher interface, which is used by any object
that fetches JSON using a fetchable property. It defines the FetchJSON
method and function (the function FetchJSON uses a default fetcher for
convenience).
This is for smoother integration with existing code and testing.

The purpose of these changes are to make unit testing easier and to
better organize code. It should also reduce repetition (DRY).